distributed_platform/clients/windows_python/requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
requests==2.32.3
websocket-client==1.8.0
insightface==0.7.3
onnxruntime-gpu==1.20.1
